package refactor for lists
